Move transformer to internals
